Move test mocks to __mocks__ folders
