+++ packages/sdk/src/__mocks__/MockAssets.ts
