+++ packages/sdk/src/lend/__mocks__/MockMarkets.ts
@@ -1,6 +1,6 @@
 import type { Address } from 'viem'
 
-import { MockUSDCAsset, MockWETHAsset } from '@/test/MockAssets.js'
+import { MockUSDCAsset, MockWETHAsset } from '@/__mocks__/MockAssets.js'
 import type { LendMarketConfig } from '@/types/lend/index.js'
 
 /**
