+++ packages/sdk/src/lend/core/__tests__/LendProvider.test.ts
@@ -1,8 +1,8 @@
 import type { Address } from 'viem'
 import { describe, expect, it } from 'vitest'
 
-import { MockUSDCAsset } from '@/test/MockAssets.js'
-import { MockLendProvider } from '@/test/MockLendProvider.js'
+import { MockUSDCAsset } from '@/__mocks__/MockAssets.js'
+import { MockLendProvider } from '@/lend/__mocks__/MockLendProvider.js'
 import type { LendMarketConfig, LendMarketId } from '@/types/lend/index.js'
 
 // Test helper class that exposes protected validation methods as public
