+++ packages/sdk/src/wallet/core/providers/hosted/registry/__tests__/HostedWalletProviderRegistry.spec.ts
@@ -2,12 +2,12 @@ import type { PrivyClient } from '@privy-io/node'
 import { unichain } from 'viem/chains'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
 import {
   createMockPrivyClient,
   getMockAuthorizationContext,
-} from '@/test/MockPrivyClient.js'
+} from '@/__mocks__/MockPrivyClient.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
+import type { ChainManager } from '@/services/ChainManager.js'
 import { HostedWalletProviderRegistry } from '@/wallet/core/providers/hosted/registry/HostedWalletProviderRegistry.js'
 import { PrivyHostedWalletProvider } from '@/wallet/node/providers/hosted/privy/PrivyHostedWalletProvider.js'
 import type { NodeOptionsMap } from '@/wallet/node/providers/hosted/types/index.js'
