+++ packages/sdk/src/wallet/core/wallets/eoa/__tests__/EOAWallet.spec.ts
@@ -10,9 +10,9 @@ import { createWalletClient } from 'viem'
 import { unichain } from 'viem/chains'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
+import { getRandomAddress } from '@/__mocks__/utils.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
-import { getRandomAddress } from '@/test/utils.js'
 import type { TransactionData } from '@/types/lend/index.js'
 import type { EOATransactionReceipt } from '@/wallet/core/wallets/abstract/types/index.js'
 import { EOAWallet } from '@/wallet/core/wallets/eoa/EOAWallet.js'
