+++ packages/sdk/src/wallet/core/wallets/smart/default/__tests__/DefaultSmartWallet.spec.ts
@@ -8,11 +8,11 @@ import { toCoinbaseSmartAccount } from 'viem/account-abstraction'
 import { baseSepolia, unichain } from 'viem/chains'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
+import { getRandomAddress } from '@/__mocks__/utils.js'
+import { createMockLendProvider } from '@/lend/__mocks__/MockLendProvider.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
 import { SUPPORTED_TOKENS } from '@/supported/tokens.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
-import { createMockLendProvider } from '@/test/MockLendProvider.js'
-import { getRandomAddress } from '@/test/utils.js'
 import type { LendProviderConfig } from '@/types/actions.js'
 import type { LendProvider, TransactionData } from '@/types/lend/index.js'
 import {
