+++ packages/sdk/src/wallet/node/wallets/hosted/turnkey/__tests__/TurnkeyWallet.spec.ts
@@ -7,9 +7,9 @@ import { createWalletClient } from 'viem'
 import { unichain } from 'viem/chains'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
+import { getRandomAddress } from '@/__mocks__/utils.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
-import { getRandomAddress } from '@/test/utils.js'
 import { TurnkeyWallet } from '@/wallet/node/wallets/hosted/turnkey/TurnkeyWallet.js'
 
 vi.mock('viem', async () => ({
