+++ packages/sdk/src/wallet/react/providers/hosted/privy/__tests__/PrivyHostedWalletProvider.spec.ts
@@ -2,9 +2,9 @@ import type { ConnectedWallet } from '@privy-io/react-auth'
 import type { LocalAccount } from 'viem'
 import { describe, expect, it, vi } from 'vitest'
 
+import { createMockLendProvider } from '@/lend/__mocks__/MockLendProvider.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
-import { createMockLendProvider } from '@/test/MockLendProvider.js'
 import { PrivyHostedWalletProvider } from '@/wallet/react/providers/hosted/privy/PrivyHostedWalletProvider.js'
 import { PrivyWallet } from '@/wallet/react/wallets/hosted/privy/PrivyWallet.js'
 import * as createSignerUtil from '@/wallet/react/wallets/hosted/privy/utils/createSigner.js'
