+++ packages/sdk/src/wallet/react/providers/registry/__tests__/ReactHostedWalletProviderRegistry.spec.ts
@@ -1,8 +1,8 @@
 import { unichain } from 'viem/chains'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
 import { DynamicHostedWalletProvider } from '@/wallet/react/providers/hosted/dynamic/DynamicHostedWalletProvider.js'
 import { PrivyHostedWalletProvider } from '@/wallet/react/providers/hosted/privy/PrivyHostedWalletProvider.js'
 import { TurnkeyHostedWalletProvider } from '@/wallet/react/providers/hosted/turnkey/TurnkeyHostedWalletProvider.js'
