+++ packages/sdk/src/wallet/react/wallets/hosted/dynamic/utils/__tests__/createSigner.spec.ts
@@ -5,7 +5,7 @@ import type { LocalAccount, WalletClient } from 'viem'
 import { toAccount } from 'viem/accounts'
 import { describe, expect, it, vi } from 'vitest'
 
-import { getRandomAddress } from '@/test/utils.js'
+import { getRandomAddress } from '@/__mocks__/utils.js'
 import { createSigner } from '@/wallet/react/wallets/hosted/dynamic/utils/createSigner.js'
 
 vi.mock('@dynamic-labs/ethereum', async () => ({
