+++ packages/sdk/src/wallet/react/wallets/hosted/privy/utils/__tests__/createSigner.spec.ts
@@ -4,7 +4,7 @@ import type { LocalAccount } from 'viem'
 import { toAccount } from 'viem/accounts'
 import { describe, expect, it, vi } from 'vitest'
 
-import { getRandomAddress } from '@/test/utils.js'
+import { getRandomAddress } from '@/__mocks__/utils.js'
 import { createSigner } from '@/wallet/react/wallets/hosted/privy/utils/createSigner.js'
 
 vi.mock('@privy-io/react-auth', async () => ({
