+++ packages/sdk/src/wallet/react/wallets/hosted/turnkey/__tests__/TurnkeyWallet.spec.ts
@@ -5,9 +5,9 @@ import { createWalletClient } from 'viem'
 import { unichain } from 'viem/chains'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
+import { getRandomAddress } from '@/__mocks__/utils.js'
+import { MockChainManager } from '@/services/__mocks__/MockChainManager.js'
 import type { ChainManager } from '@/services/ChainManager.js'
-import { MockChainManager } from '@/test/MockChainManager.js'
-import { getRandomAddress } from '@/test/utils.js'
 import { TurnkeyWallet } from '@/wallet/react/wallets/hosted/turnkey/TurnkeyWallet.js'
 
 vi.mock('viem', async () => ({
