+++ packages/sdk/src/wallet/react/wallets/hosted/turnkey/utils/__tests__/createSigner.spec.ts
@@ -3,7 +3,7 @@ import { createAccount } from '@turnkey/viem'
 import type { LocalAccount } from 'viem'
 import { describe, expect, it, vi } from 'vitest'
 
-import { getRandomAddress } from '@/test/utils.js'
+import { getRandomAddress } from '@/__mocks__/utils.js'
 import { createSigner } from '@/wallet/react/wallets/hosted/turnkey/utils/createSigner.js'
 
 vi.mock('@turnkey/viem', async () => ({
